Refactoring UI Package
Refactored the MealSwapUI by addressing Duplicate Code in the initialize method by creating methods for the duplicates such as addSeparator for the gridBagLayout.

Refactored applyTooltipAndColoring() method in the MealSwapUI class to address the Deep Nesting code smell. This is done by creating two new helpet methods createdStyledCell() and applyColorAndTooltip() for the lambda function in the original method.

Lastly, addressed the Data Clumps code smell and refactored the constructor of MealEntry class by condensing the parameters to a new class called NutrientProfile which takes in calories, protein, carbs and fat. This reduces the load of parameters for MealEntry.